Renamed headers and source files from 'alps_history.h' to 'alps/history.h'.
