+++ examples/evo_real/Makefile.in
@@ -188,7 +188,7 @@ target_alias = @target_alias@
 top_build_prefix = @top_build_prefix@
 top_builddir = @top_builddir@
 top_srcdir = @top_srcdir@
-AM_CPPFLAGS = -g -Wall -I../../src -I../../include/alps
+AM_CPPFLAGS = -g -Wall -I../../src -I../../include
 evo_real_SOURCES = evo_real.cpp
 evo_real_LDADD = ../../src/libalps.la
 all: all-am
