+++ examples/evo_real/evo_real.cpp
@@ -21,19 +21,19 @@
 
 ********************************************************/
 
-#include "alps_random_mt.h"
+#include "alps/random_mt.h"
 using namespace alps_random;
 
 
-#include "alps_utils.h"
+#include "alps/utils.h"
 using namespace alps_utils;
 
 
-#include "alps.h"
-#include "alps_gen.h"
-#include "alps_sstate.h"
-#include "alps_layer.h"
-#include "alps_individ_real.h"
+#include "alps/alps.h"
+#include "alps/gen.h"
+#include "alps/sstate.h"
+#include "alps/layer.h"
+#include "alps/individ_real.h"
 using namespace alps;
