+++ examples/evo_real_barebones/Makefile.am
@@ -1,4 +1,4 @@
-AM_CPPFLAGS = -Wall -Wextra -ansi -pedantic -I../../src -I../../include/alps
+AM_CPPFLAGS = -Wall -Wextra -ansi -pedantic -I../../src -I../../include
 #CFLAGS = -O3 -Wall -finline-functions -fforce-addr -fomit-frame-pointer -fexpensive-optimizations -funroll-loops -frerun-cse-after-loop -frerun-loop-opt -falign-functions=4 
 
 # We want to build evoreal but we don't want to install it.
