+++ include/alps/Makefile.am
@@ -1,0 +1,3 @@
+libalps_la_includedir = $(includedir)/alps
+libalps_la_include_HEADERS = alps.h gen.h history.h individ.h individ_bits.h individ_real.h layer.h random_mt.h sstate.h utils.h
+
