+++ include/alps/individ_bits.h
