+++ include/alps/sstate.h
