+++ src/Makefile.in
@@ -74,9 +74,8 @@ am__base_list = \
 am__installdirs = "$(DESTDIR)$(libdir)"
 LTLIBRARIES = $(lib_LTLIBRARIES)
 libalps_la_LIBADD =
-am_libalps_la_OBJECTS = alps.lo alps_gen.lo alps_history.lo \
-	alps_individ.lo alps_individ_real.lo alps_layer.lo \
-	alps_random_mt.lo alps_sstate.lo alps_utils.lo
+am_libalps_la_OBJECTS = alps.lo gen.lo history.lo individ.lo \
+	individ_real.lo layer.lo random_mt.lo sstate.lo utils.lo
 libalps_la_OBJECTS = $(am_libalps_la_OBJECTS)
 libalps_la_LINK = $(LIBTOOL) --tag=CXX $(AM_LIBTOOLFLAGS) \
 	$(LIBTOOLFLAGS) --mode=link $(CXXLD) $(AM_CXXFLAGS) \
@@ -217,11 +216,11 @@ top_builddir = @top_builddir@
 top_srcdir = @top_srcdir@
 
 #AM_CPPFLAGS = -O3 -Wall -finline-functions -fforce-addr -fomit-frame-pointer -fexpensive-optimizations -funroll-loops -frerun-cse-after-loop -frerun-loop-opt -falign-functions=4 
-AM_CPPFLAGS = -Wall -Wextra -ansi -pedantic -I../include/alps
+AM_CPPFLAGS = -Wall -Wextra -ansi -pedantic -I../include
 lib_LTLIBRARIES = libalps.la
 
 # The alps_individ_bits class not included; it breaks the build.
-libalps_la_SOURCES = alps.cpp alps_gen.cpp alps_history.cpp alps_individ.cpp alps_individ_real.cpp alps_layer.cpp alps_random_mt.cpp alps_sstate.cpp alps_utils.cpp
+libalps_la_SOURCES = alps.cpp gen.cpp history.cpp individ.cpp individ_real.cpp layer.cpp random_mt.cpp sstate.cpp utils.cpp
 
 # There are two library versioning systems.  One aligns with the
 # release number.  The other is the version-info which has a trickier
@@ -365,14 +364,14 @@ distclean-compile:
 	-rm -f *.tab.c
 
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/alps.Plo@am__quote@
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/alps_gen.Plo@am__quote@
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/alps_history.Plo@am__quote@
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/alps_individ.Plo@am__quote@
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/alps_individ_real.Plo@am__quote@
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/alps_layer.Plo@am__quote@
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/alps_random_mt.Plo@am__quote@
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/alps_sstate.Plo@am__quote@
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/alps_utils.Plo@am__quote@
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/gen.Plo@am__quote@
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/history.Plo@am__quote@
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/individ.Plo@am__quote@
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/individ_real.Plo@am__quote@
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/layer.Plo@am__quote@
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/random_mt.Plo@am__quote@
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/sstate.Plo@am__quote@
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/utils.Plo@am__quote@
 
 .cpp.o:
 @am__fastdepCXX_TRUE@	$(CXXCOMPILE) -MT $@ -MD -MP -MF $(DEPDIR)/$*.Tpo -c -o $@ $<
