+++ src/sstate.cpp
@@ -30,17 +30,17 @@
 #include <cstdlib>
 
 
-#include "alps_random_mt.h"
+#include "alps/random_mt.h"
 using namespace alps_random;
 
-#include "alps_utils.h"
+#include "alps/utils.h"
 using namespace alps_utils;
 
-#include "alps.h"
-#include "alps_sstate.h"
-#include "alps_layer.h"
-#include "alps_individ.h"
-#include "alps_history.h"
+#include "alps/alps.h"
+#include "alps/sstate.h"
+#include "alps/layer.h"
+#include "alps/individ.h"
+#include "alps/history.h"
 
 using namespace std;
